add compatibility with the previous version

+++ .warp/setup/init/info.sh
@@ -4,13 +4,13 @@ warp_message "* Configuring environment variable files $(warp_message_ok [ok])"
  [ ! -f $DOCKERIGNOREFILE ] && cp $PROJECTPATH/.warp/setup/init/.dockerignore $DOCKERIGNOREFILE
 
 # creating ext-xdebug.ini
-if  [ ! -f $PROJECTPATH/.warp/docker/config/php/ext-xdebug.ini ]
+if  [ ! -f $PROJECTPATH/.warp/docker/config/php/ext-xdebug.ini ] && [ -f $PROJECTPATH/.warp/docker/config/php/ext-xdebug.ini.sample ]
 then
     cp $PROJECTPATH/.warp/docker/config/php/ext-xdebug.ini.sample $PROJECTPATH/.warp/docker/config/php/ext-xdebug.ini
 fi
 
 # creating ext-ioncube.ini
-if  [ ! -f $PROJECTPATH/.warp/docker/config/php/ext-ioncube.ini ]
+if  [ ! -f $PROJECTPATH/.warp/docker/config/php/ext-ioncube.ini ] && [ -f $PROJECTPATH/.warp/docker/config/php/ext-ioncube.ini.sample ]
 then
     cp $PROJECTPATH/.warp/docker/config/php/ext-ioncube.ini.sample $PROJECTPATH/.warp/docker/config/php/ext-ioncube.ini
 fi
